refactor(context/execute): extract signing submodule from the execute god-file
First increment of splitting the 2675-line execute handler. Moves the
authorized-action signing/persistence helpers into `signing.rs` —
behavior-preserving, pure code movement (mod.rs: +2 wiring lines, -300).

Moved (free functions):
- sign_authorized_actions, persist_signed_signatures

Both are `pub(crate)` and imported by `handlers::create_context` via
`crate::handlers::execute::{..}`; re-exported `pub(crate)` from the execute
module so that path is unchanged. Explicit imports, no `use super::*`.

Validated: context lib tests (107) pass; cargo check clean.
